Refactor page headers and table header row usage
