+++ extension/devtools/app/lib/features/collections_page.dart
@@ -13,7 +13,6 @@ import '../shared/widgets/inline_empty_state.dart';
 import '../shared/widgets/page_header.dart';
 import '../shared/widgets/panel.dart';
 import '../shared/widgets/sort_header_cell.dart';
-import '../shared/widgets/table_header_row.dart';
 
 class CollectionsPage extends StatelessWidget {
   const CollectionsPage({super.key});
@@ -263,7 +262,7 @@ class _CollectionsHeaderRow extends StatelessWidget {
   Widget build(BuildContext context) {
     final labelStyle = Theme.of(context).textTheme.labelSmall;
 
-    return TableHeaderRow(
+    return _TableHeaderRow(
       child: Row(
         children: [
           Expanded(
@@ -309,6 +308,27 @@ class _CollectionsHeaderRow extends StatelessWidget {
           ),
         ],
       ),
+    );
+  }
+}
+
+class _TableHeaderRow extends StatelessWidget {
+  const _TableHeaderRow({required this.child});
+
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+      decoration: BoxDecoration(
+        border: Border(
+          bottom: BorderSide(
+            color: Theme.of(context).dividerColor.withValues(alpha: 0.4),
+          ),
+        ),
+      ),
+      child: child,
     );
   }
 }
